Fixing sacss failure introduced in changelist 183045.  alignments are unpacked from varargs as promoted to size_t, but mps_pf_align is a naked number literal, and so was passed as int, causing a corrupt alignemnt failure in sacss.
Copied from Perforce
 Change: 183048
 ServerID: perforce.ravenbrook.com

+++ mps/code/sacss.c
@@ -173,7 +173,7 @@ static int testInArena(mps_arena_t arena)
 
   printf("MVFF\n\n");
   die(stress(mps_class_mvff(), classCOUNT, classes, randomSize8, arena,
-             (size_t)65536, (size_t)32, MPS_PF_ALIGN, TRUE, TRUE, TRUE),
+             (size_t)65536, (size_t)32, (size_t)MPS_PF_ALIGN, TRUE, TRUE, TRUE),
       "stress MVFF");
   printf("MV debug\n\n");
   die(stress(mps_class_mv_debug(), classCOUNT, classes, randomSize8, arena,
